MoticCamera: Prevent VC++ x64 warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10948 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Motic/MoticCamera.cpp
@@ -983,7 +983,7 @@ int CMoticCamera::OnDevice( MM::PropertyBase* pProp, MM::ActionType eAct )
     if (it != end && idx != m_iCurDeviceIdx)
     {
       StopSequenceAcquisition();
-      if(MIDP_SelectCamera(idx) != 0)
+      if(MIDP_SelectCamera(static_cast<long>(idx)) != 0)
         return DEVICE_ERR;
       m_iCurDeviceIdx = MIDP_GetCurCameraIndex();
       InitDevice();
